Created a seperate controller and model for article, Updated controller and model of topic, Updated app.js to require different controllers)
